requirements.txt
streamlit
google-genai
pymupdf==1.23.8
python-docx
